# docs/swagger.yaml
definitions:
  main.AddRevokeCertRequest:
    properties:
      expiration_date:
        example: "2025-12-31T23:59:59Z"
        type: string
      issuer_key_hash:
        example: 10d8ff2cf856bac45cb80e8fb83a566cd3535d93
        type: string
      revocation_date:
        example: "2025-05-01T12:00:00Z"
        type: string
      revocation_reason:
        example: "1"
        type: string
      serial_number:
        example: "1234"
        type: string
    type: object
  main.ListResponderCertsResponse:
    properties:
      certificates:
        items:
          type: string
        type: array
    type: object
  main.ListRevokedCertsResponse:
    properties:
      revoked_certs:
        additionalProperties:
          additionalProperties:
            $ref: '#/definitions/main.OCSPEntry'
          type: object
        type: object
    type: object
  main.OCSPEntry:
    properties:
      expiration_date:
        description: Certificate expiration date
        type: string
      revocation_date:
        description: Revocation date
        type: string
      revocation_reason:
        description: Revocation reason
        type: string
      serial_number:
        description: Serial number (hex)
        type: string
    type: object
  main.RemoveResponderRequest:
    properties:
      issuer_certificate:
        example: |-
          -----BEGIN CERTIFICATE-----
          MIID...AB
          -----END CERTIFICATE-----
        type: string
      ocsp_certificate:
        example: |-
          -----BEGIN CERTIFICATE-----
          MIIF...AB
          -----END CERTIFICATE-----
        type: string
    type: object
  main.RemoveRevokeCertRequest:
    properties:
      issuer_key_hash:
        example: 10d8ff2cf856bac45cb80e8fb83a566cd3535d93
        type: string
      serial_number:
        example: "1234"
        type: string
    type: object
  main.UploadSignedCertRequest:
    properties:
      issuer_certificate:
        example: |-
          -----BEGIN CERTIFICATE-----
          MIIF...CD
          -----END CERTIFICATE-----
        type: string
      signed_certificate:
        example: |-
          -----BEGIN CERTIFICATE-----
          MIID...AB
          -----END CERTIFICATE-----
        type: string
    type: object
  main.createNewCsrRequest:
    properties:
      common_name:
        example: simple-va.de
        type: string
    type: object
  main.createNewCsrResponse:
    properties:
      csr:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1.0.0/addrevokedcert:
    post:
      consumes:
      - application/json
      description: Marks a certificate as revoked using issuer key hash, serial number,
        and revocation metadata.
      parameters:
      - description: Certificate revocation details
        in: body
        name: cert
        required: true
        schema:
          $ref: '#/definitions/main.AddRevokeCertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Certificate successfully revoked
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Failed to revoke certificate
          schema:
            type: string
      summary: Add a revoked certificate
      tags:
      - RevokedCertsAction
  /v1.0.0/createnewcsr:
    post:
      consumes:
      - application/json
      description: Generates a new ECDSA CSR for the given common name.
      parameters:
      - description: Common name for the CSR
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.createNewCsrRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.createNewCsrResponse'
        "400":
          description: CommonName is required or CSR generation failed
          schema:
            type: string
      summary: Create a new Certificate Signing Request (CSR)
      tags:
      - ResponderCertsAction
  /v1.0.0/listrespondercerts:
    get:
      description: Retrieves all OCSP responder certificates stored in the manager,
        returned as PEM-encoded strings.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ListResponderCertsResponse'
      summary: List all OCSP certificates
      tags:
      - ResponderCertsAction
  /v1.0.0/listrevokedcerts:
    get:
      description: Retrieves all revoked certificates
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ListRevokedCertsResponse'
      summary: List all revoked certificates
      tags:
      - RevokedCertsAction
  /v1.0.0/removeresponder:
    post:
      consumes:
      - application/json
      description: Removes an OCSP responder identified by its issuer certificate
        and responder certificate (both PEM-encoded).
      parameters:
      - description: Issuer cert and OCSP cert in PEM format
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RemoveResponderRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Certificate successfully removed
          schema:
            type: string
        "400":
          description: Bad request (e.g. invalid JSON)
          schema:
            type: string
        "500":
          description: Failed to remove certificate
          schema:
            type: string
      summary: Remove an OCSP responder
      tags:
      - ResponderCertsAction
  /v1.0.0/removerevokedcert:
    post:
      consumes:
      - application/json
      description: Remove a revoked certificate from the list
      parameters:
      - description: Certificate revocation removal details
        in: body
        name: cert
        required: true
        schema:
          $ref: '#/definitions/main.RemoveRevokeCertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Certificate successfully removed
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Failed to remove certificate
          schema:
            type: string
      summary: Remove a revoked certificate from the list
      tags:
      - RevokedCertsAction
  /v1.0.0/uploadsignedcert:
    post:
      consumes:
      - application/json
      description: Uploads a signed OCSP responder certificate along with its issuer
        certificate (both PEM-encoded).
      parameters:
      - description: Signed OCSP cert and issuer cert in PEM format
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.UploadSignedCertRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Certificate uploaded successfully
          schema:
            type: string
        "400":
          description: Bad request (e.g. missing fields)
          schema:
            type: string
        "500":
          description: Failed to upload certificate
          schema:
            type: string
      summary: Upload a signed OCSP responder certificate
      tags:
      - ResponderCertsAction
swagger: "2.0"